# contracts/oar-openapi.yaml
openapi: 3.1.0
info:
  title: Organization Autorunner HTTP API
  version: 0.2.3
  description: |
    Canonical HTTP contract for organization-autorunner.

    This OpenAPI contract is the monorepo source of truth for generated clients,
    CLI command registry metadata, and command docs.
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /threads:
    post:
      operationId: createThread
      summary: Create thread snapshot
      tags: [threads]
      x-oar-surface: "canonical"
      x-oar-command-id: "threads.create"
      x-oar-cli-path: "threads create"
      x-oar-why: "Open a new thread for tracking ongoing organizational work."
      x-oar-examples:
        - title: Create thread
          command: oar threads create --from-file thread.json --json
      x-oar-input-mode: "json-body"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ thread }` including generated id and audit fields."
      x-oar-error-codes: [invalid_json, invalid_request, unknown_actor_id, conflict]
      x-oar-concepts: [threads, snapshots]
      x-oar-stability: "stable"
      x-oar-agent-notes: "Replay-safe when `request_key` is reused with the same body; otherwise core issues a new canonical thread id."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThreadRequest'
      responses:
        '201':
          description: Thread created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    get:
      operationId: listThreads
      summary: List thread snapshots
      tags: [threads]
      x-oar-surface: "canonical"
      x-oar-command-id: "threads.list"
      x-oar-cli-path: "threads list"
      x-oar-why: "Retrieve current thread state for triage and scheduling decisions."
      x-oar-examples:
        - title: List active p1 threads
          command: oar threads list --status active --priority p1 --json
        - title: Search threads by title
          command: oar threads list --q "launch" --json
        - title: Paginated thread list
          command: oar threads list --limit 20 --json
      x-oar-input-mode: "none"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ threads, next_cursor? }`; query filters are additive. Pagination is optional and backward-compatible."
      x-oar-error-codes: [invalid_request]
      x-oar-concepts: [threads, filtering]
      x-oar-stability: "stable"
      x-oar-agent-notes: "Safe and idempotent. Optional pagination with `q` for search, `limit` for page size, and `cursor` for continuation."
      parameters:
        - in: query
          name: q
          schema: { type: string }
          description: "Optional search query to filter threads by ID or title using case-insensitive matching."
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000 }
          description: "Maximum number of results to return. Defaults to returning all results when omitted."
        - in: query
          name: cursor
          schema: { type: string }
          description: "Opaque pagination cursor from a previous response's `next_cursor` field."
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: priority
          schema: { type: string }
        - in: query
          name: tag
          schema: { type: array, items: { type: string } }
          style: form
          explode: true
        - in: query
          name: cadence
          schema: { type: array, items: { type: string } }
          style: form
          explode: true
        - in: query
          name: stale
          schema: { type: boolean }
      responses:
        '200':
          description: Thread list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadsResponse'
...
  /commitments:
    post:
      operationId: createCommitment
      summary: Create commitment snapshot
      tags: [commitments]
      x-oar-surface: "canonical"
      x-oar-command-id: "commitments.create"
      x-oar-cli-path: "commitments create"
      x-oar-why: "Track accountable work items tied to a thread."
      x-oar-examples:
        - title: Create commitment
          command: oar commitments create --from-file commitment.json --json
      x-oar-input-mode: "json-body"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ commitment }` with generated id."
      x-oar-error-codes: [invalid_json, invalid_request, unknown_actor_id, conflict]
      x-oar-concepts: [commitments]
      x-oar-stability: "stable"
      x-oar-agent-notes: "Replay-safe when `request_key` is reused with the same body; otherwise each create issues a new commitment id."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommitmentRequest'
      responses:
        '201':
          description: Commitment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitmentResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    get:
      operationId: listCommitments
      summary: List commitments
      tags: [commitments]
      x-oar-surface: "canonical"
      x-oar-command-id: "commitments.list"
      x-oar-cli-path: "commitments list"
      x-oar-why: "Monitor open/blocked work and due windows."
      x-oar-examples:
        - title: List open commitments for a thread
          command: oar commitments list --thread-id thread_123 --status open --json
      x-oar-input-mode: "none"
      x-oar-streaming: { mode: none }
      x-oar-output-envelope: "Returns `{ commitments }`."
      x-oar-error-codes: [invalid_request]
      x-oar-concepts: [commitments, filtering]
      x-oar-stability: "stable"
      x-oar-agent-notes: "Safe and idempotent."
      parameters:
        - in: query
          name: thread_id
          schema: { type: string }
        - in: query
          name: owner
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: due_before
          schema: { type: string, format: date-time }
        - in: query
          name: due_after
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Commitment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitmentsResponse'
...
components:
...
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            recoverable: { type: boolean }
            hint: { type: string }
          required: [code, message, recoverable, hint]
      required: [error]
...
    Thread:
      type: object
      additionalProperties: true
    Commitment:
      type: object
      additionalProperties: true
...
    CreateThreadRequest:
      type: object
      properties:
        actor_id: { type: string }
        request_key: { type: string }
        thread:
          $ref: '#/components/schemas/AnyObject'
      required: [thread]
...
    ThreadResponse:
      type: object
      properties:
        thread:
          $ref: '#/components/schemas/Thread'
      required: [thread]
    ThreadsResponse:
      type: object
      properties:
        threads:
          type: array
          items:
            $ref: '#/components/schemas/Thread'
        next_cursor:
          type: string
          description: "Opaque pagination cursor for fetching the next page of results. Omitted when there are no more results."
      required: [threads]
...
    CreateCommitmentRequest:
      type: object
      properties:
        actor_id: { type: string }
        request_key: { type: string }
        commitment:
          $ref: '#/components/schemas/AnyObject'
      required: [commitment]
...
    CommitmentResponse:
      type: object
      properties:
        commitment:
          $ref: '#/components/schemas/Commitment'
      required: [commitment]
    CommitmentsResponse:
      type: object
      properties:
        commitments:
          type: array
          items:
            $ref: '#/components/schemas/Commitment'
      required: [commitments]
...